+++ CMS/ASN1/PKCS7/EncryptedData.h
@@ -33,5 +33,5 @@ EncryptedData ::= SEQUENCE {
 
 				void	Set(ObjectIdentifier*		_type,
 							Encryption*				_algorithm,
-							unsigned	int			_szContent);
+							size_t					_szContent);
 };
